Dia 05: refatorando constantes dos testes
