+++ tests/aux/helpers.js
@@ -1,0 +1,7 @@
+const getGameBoardLetter = (index) => {
+    return global.document.querySelector(`.letter-${index}`).style.backgroundColor
+}
+
+module.exports = {
+    getGameBoardLetter
+}
